move getData logic to measurements model
