+++ lib/controllers/measurementsController.js
@@ -4,14 +4,11 @@ const
   restifyErrors = require('restify-errors'),
   { Measurement, Box } = require('../models'),
   csvstringify = require('csv-stringify'),
-  streamTransform = require('stream-transform'),
-  jsonstringify = require('stringify-stream'),
   { checkContentType } = require('../helpers/apiUtils'),
   {
     retrieveParameters,
     validateFromToTimeParams
   } = require('../helpers/userParamHelpers'),
-  { outlierTransformer } = require('../statistics'),
   handleError = require('../helpers/errorHandler');
 
 const { BadRequestError, UnsupportedMediaTypeError } = restifyErrors;
@@ -31,12 +28,6 @@ const getLatestMeasurements = function getLatestMeasurements (req, res, next) {
     .catch(function (err) {
       handleError(err, next);
     });
-};
-
-const getDataTransformerFunction = function getDataTransformerFunction (data) {
-  data.createdAt = data.createdAt.toISOString();
-
-  return data;
 };
 
 /**
@@ -60,70 +51,23 @@ const getDataTransformerFunction = function getDataTransformerFunction (data) {
  * @apiUse SeparatorParam
  */
 const getData = function getData (req, res, next) {
-  let stringifier;
-
-  const { format, delimiter, outliers, outlierWindow } = req._userParams;
-
-  // validate window
-  if (outlierWindow < 1 || outlierWindow >= 50) {
-    return next(new BadRequestError('parameter outlier-window must be between 1 and 50, default is 15'));
-  }
+  const { sensorId, format, download } = req._userParams;
 
   // IDEA: add geojson point featurecollection format
-  if (format === 'csv') {
+  if (format === 'csv' || (download === 'true')) {
     res.header('Content-Type', 'text/csv');
-    const csvColumns = ['createdAt', 'value'];
-    if (outliers) {
-      csvColumns.push('isOutlier');
-    }
-
-    stringifier = csvstringify({ columns: csvColumns, header: 1, delimiter });
+    res.header('Content-Disposition', `attachment; filename=${sensorId}.${format}`);
   } else if (format === 'json') {
     res.header('Content-Type', 'application/json; charset=utf-8');
-    stringifier = jsonstringify({ open: '[', close: ']' }, function replacer (k, v) {
-      // dont send unnecessary nested location
-      return (k === 'location') ? v.coordinates : v;
-    });
   }
 
-  // offer download to browser
-  if (format === 'csv' || (req._userParams.download === 'true')) {
-    res.header('Content-Disposition', `attachment; filename=${req.params.sensorId}.${format}`);
-  }
+  Measurement.getMeasurementsStream(req._userParams)
+    .on('error', function (err) {
+      res.end(`Error: ${err.message}`);
 
-  // finally execute the query
-  const queryLimit = 10000;
-
-  const qry = {
-    sensor_id: req._userParams.sensorId,
-    createdAt: {
-      $gte: req._userParams.fromDate.toDate(),
-      $lte: req._userParams.toDate.toDate()
-    }
-  };
-
-  const measurementsCursor = Measurement
-    .find(qry, { 'createdAt': 1, 'value': 1, 'location': 1, '_id': 0 })
-    .cursor({ lean: true, limit: queryLimit });
-
-  if (outliers) {
-    measurementsCursor
-      .pipe(streamTransform(getDataTransformerFunction))
-      .pipe(outlierTransformer({
-        window: Math.trunc(outlierWindow), // only allow integer values
-        replaceOutlier: (outliers === 'replace')
-      }))
-      .on('error', function (err) {
-        res.end(`Error: ${err.message}`);
-      })
-      .pipe(stringifier)
-      .pipe(res);
-  } else {
-    measurementsCursor
-      .pipe(streamTransform(getDataTransformerFunction))
-      .pipe(stringifier)
-      .pipe(res);
-  }
+      return next(err);
+    })
+    .pipe(res);
 };
 
 /**
