+++ lib/models/measurement.js
@@ -4,7 +4,11 @@ const { mongoose } = require('../db'),
   Schema = mongoose.Schema,
   moment = require('moment'),
   decodeHandlers = require('../decoding'),
-  ModelError = require('./modelError');
+  ModelError = require('./modelError'),
+  csvstringify = require('csv-stringify'),
+  streamTransform = require('stream-transform'),
+  jsonstringify = require('stringify-stream'),
+  { outlierTransformer } = require('../statistics');
 
 const measurementSchema = new Schema({
   value: {
@@ -121,6 +125,57 @@ measurementSchema.statics.findLatestMeasurementsForSensors = function findLatest
     { $project: { value: 1, createdAt: 1, sensor_id: 1, _id: 0 } }
   ])
     .exec();
+};
+
+const csvColumns = ['createdAt', 'value'];
+
+const jsonLocationReplacer = function jsonLocationReplacer (k, v) {
+  // dont send unnecessary nested location
+  return (k === 'location') ? v.coordinates : v;
+};
+
+const getDataTransformerFunction = function getDataTransformerFunction (data) {
+  data.createdAt = data.createdAt.toISOString();
+
+  return data;
+};
+
+measurementSchema.statics.getMeasurementsStream = function getMeasurementsStream ({ format, delimiter, fromDate, toDate, sensorId, outliers, outlierWindow }) {
+  let stringifier;
+  // IDEA: add geojson point featurecollection format
+  if (format === 'csv') {
+    stringifier = csvstringify({ columns: csvColumns, header: 1, delimiter });
+  } else if (format === 'json') {
+    stringifier = jsonstringify({ open: '[', close: ']' }, jsonLocationReplacer);
+  }
+
+  // finally execute the query
+  const queryLimit = 10000;
+
+  const qry = {
+    sensor_id: sensorId,
+    createdAt: {
+      $gte: fromDate.toDate(),
+      $lte: toDate.toDate()
+    }
+  };
+
+  let measurementsCursor = this
+    .find(qry, { 'createdAt': 1, 'value': 1, 'location': 1, '_id': 0 })
+    .cursor({ lean: true, limit: queryLimit })
+    .pipe(streamTransform(getDataTransformerFunction));
+
+  if (outliers) {
+    measurementsCursor = measurementsCursor
+      .pipe(outlierTransformer({
+        window: Math.trunc(outlierWindow), // only allow integer values
+        replaceOutlier: (outliers === 'replace')
+      }))
+      .pipe(stringifier);
+  }
+
+  return measurementsCursor
+    .pipe(stringifier);
 };
 
 const measurementModel = mongoose.model('Measurement', measurementSchema);
